Drop rubocop-minitest from gem dependency
Closes #13.

This PR drops rubocop-minitest from gem dependency.

In fact, the `Minitest` department's Cop is not used in rubocop-rails-omakase:
https://github.com/rails/rubocop-rails-omakase/blob/v1.0.0/rubocop.yml

This means that it might be appropriate for users to add the `Minitest` department's cop only when they start using it.
As a result, there will be no need to install gems that are not included in the rubocop-rails-omakase configuration.

Users can be given the choice to add cops based on the testing framework they are using.